the backend is almost complete and frontend is started

+++ src/main/java/com/bankingapp/banksystem/factory/TransactionFactory.java
@@ -5,5 +5,5 @@ import com.bankingapp.banksystem.model.Transaction;
 import java.math.BigDecimal;
 
 public interface TransactionFactory {
-    Transaction getTransaction(String accountType, String actionType, String obj, Double amount, BigDecimal balance);
+    Transaction getTransaction(String accountType, String actionType, String obj, BigDecimal amount, BigDecimal balance);
 }
